[TAN-1106] Move all user group related code to one concern
